email: extract SendEmailAttachment
